@@ -2,11 +2,19 @@
   "name": "webdav",
   "version": "4.11.2",
   "description": "WebDAV client for NodeJS",
-  "main": "dist/node/index.js",
+  "exports": {
+    ".": {
+      "browser": "./dist/web/index.js",
+      "default": "./dist/node/index.js"
+    },
+    "./web": "./dist/web/index.js"
+  },
+  "type": "module",
+  "types": "./dist/node/index.d.ts",
   "scripts": {
     "build": "npm run clean && run-p build:node build:web",
     "build:node": "tsc",
-    "build:web": "webpack --mode production --config webpack.config.js && tsc --emitDeclarationOnly --outDir dist/web/",
+    "build:web": "webpack --mode production --config webpack.config.cjs",
     "clean": "rimraf dist web",
     "dev": "nodemon -e ts,js --exec 'npm run build' --ignore 'dist/' --ignore 'web/'",
     "format": "prettier --write '{source,test}/**/*.{js,ts}'",
@@ -14,16 +22,14 @@
     "prepublishOnly": "run-s build prepare:publish:web",
     "test": "run-s test:node test:web test:format",
     "test:format": "prettier-check '{source,test}/**/*.{js,ts}'",
-    "test:node": "npm run build:node && nyc mocha -r test/index.node.js \"test/node/**/*.spec.js\"",
+    "test:node": "nyc mocha \"test/node/**/*.spec.ts\"",
     "test:node:watch": "nodemon --exec 'npm run test:node' --ignore 'dist/'",
     "test:web": "concurrently --success 'first' --kill-others 'npm run test:web:karma' 'npm run test:web:server'",
-    "test:web:karma": "karma start test/karma.conf.js",
-    "test:web:server": "node test/server.web.js"
+    "test:web:karma": "karma start test/karma.conf.cjs",
+    "test:web:server": "ts-node-esm ./test/server.web.ts"
   },
-  "types": "dist/node/index.d.ts",
   "files": [
     "dist/",
-    "web/",
     "test/server/",
     "*.md"
   ],
@@ -38,7 +44,7 @@
     "sync"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">=14"
   },
   "lint-staged": {
     "{source,test}/**/*.{js,ts}": [
@@ -57,12 +63,12 @@
   },
   "homepage": "https://github.com/perry-mitchell/webdav-client#readme",
   "dependencies": {
-    "axios": "^0.27.2",
     "base-64": "^1.0.0",
     "byte-length": "^1.0.2",
+    "cross-fetch": "^3.1.5",
     "fast-xml-parser": "^3.19.0",
     "he": "^1.2.0",
-    "hot-patcher": "^1.0.0",
+    "hot-patcher": "^2.0.0",
     "layerr": "^0.1.2",
     "md5": "^2.3.0",
     "minimatch": "^5.1.0",
@@ -75,13 +81,16 @@
     "@babel/core": "^7.18.10",
     "@babel/preset-env": "^7.18.10",
     "@babel/preset-typescript": "^7.18.6",
-    "@types/node": "^18.7.8",
+    "@types/chai": "^4.3.4",
+    "@types/he": "^1.1.2",
+    "@types/mocha": "^10.0.0",
+    "@types/node": "^18.11.9",
+    "@types/sinon": "^10.0.13",
     "@types/url-parse": "^1.4.8",
     "babel-loader": "^8.2.5",
     "babel-plugin-transform-async-to-promises": "^0.8.18",
     "buffer-equals": "^2.0.0",
-    "chai": "^4.3.6",
-    "chai-as-promised": "^7.1.1",
+    "chai": "^4.3.7",
     "concurrently": "^7.3.0",
     "copy-dir": "^1.3.0",
     "directory-exists": "^2.0.1",
@@ -97,19 +106,23 @@
     "karma-webpack": "^5.0.0",
     "lint-staged": "^12.5.0",
     "mkdirp": "^1.0.4",
-    "mocha": "^9.2.2",
+    "mocha": "^10.1.0",
     "nock": "^13.2.9",
+    "node-fetch": "^3.3.0",
     "nodemon": "^2.0.19",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
     "prettier": "^2.7.1",
     "prettier-check": "^2.0.0",
+    "resolve-typescript-plugin": "^2.0.0",
     "rimraf": "^3.0.2",
-    "sinon": "^13.0.2",
-    "typescript": "^4.7.4",
+    "sinon": "^14.0.2",
+    "speed-measure-webpack-plugin": "^1.5.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.3",
     "wait-on": "^6.0.1",
     "webdav-server": "^2.6.2",
-    "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0"
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.0"
   }
 }
