@@ -1,13 +1,14 @@
-const nock = require("nock");
-const { AuthType } = require("../../dist/node/index.js");
-const {
+import nock from "nock";
+import { expect } from "chai";
+import { AuthType } from "../../source/index.js";
+import {
     SERVER_PASSWORD,
     SERVER_PORT,
     SERVER_USERNAME,
     clean,
     createWebDAVClient,
     createWebDAVServer
-} = require("../helpers.node.js");
+} from "../helpers.node.js";
 
 const DUMMYSERVER = "https://dummy.webdav.server";
 
@@ -31,7 +32,7 @@ describe("auth", function () {
         nock(DUMMYSERVER)
             .get("/file")
             .reply(200, function () {
-                expect(this.req.headers.authorization).to.equal("Basic dXNlcjpwYXNz");
+                expect(this.req.headers.authorization).to.deep.equal(["Basic dXNlcjpwYXNz"]);
                 return "";
             });
         const webdav = createWebDAVClient(DUMMYSERVER, {
@@ -45,7 +46,7 @@ describe("auth", function () {
         nock(DUMMYSERVER)
             .get("/file")
             .reply(200, function () {
-                expect(this.req.headers.authorization).to.deep.equal("Bearer ABC123");
+                expect(this.req.headers.authorization).to.deep.equal(["Bearer ABC123"]);
                 return "";
             });
         const webdav = createWebDAVClient(DUMMYSERVER, {
