@@ -1,6 +1,8 @@
-const path = require("path");
-const directoryExists = require("directory-exists").sync;
-const {
+import path from "path";
+import { fileURLToPath } from "url";
+import directoryExists from "directory-exists";
+import { expect } from "chai";
+import {
     SERVER_PASSWORD,
     SERVER_PORT,
     SERVER_USERNAME,
@@ -9,7 +11,9 @@ const {
     createWebDAVServer,
     restoreRequests,
     useRequestSpy
-} = require("../../helpers.node.js");
+} from "../../helpers.node.js";
+
+const dirname = path.dirname(fileURLToPath(import.meta.url));
 
 describe("createDirectory", function () {
     beforeEach(function () {
@@ -29,10 +33,10 @@ describe("createDirectory", function () {
     });
 
     it("creates directories", async function () {
-        const newDir = path.resolve(__dirname, "../../testContents/sub2");
-        expect(directoryExists(newDir)).to.be.false;
+        const newDir = path.resolve(dirname, "../../testContents/sub2");
+        expect(directoryExists.sync(newDir)).to.be.false;
         await this.client.createDirectory("/sub2");
-        expect(directoryExists(newDir)).to.be.true;
+        expect(directoryExists.sync(newDir)).to.be.true;
     });
 
     it("allows specifying custom headers", async function () {
@@ -41,29 +45,29 @@ describe("createDirectory", function () {
                 "X-test": "test"
             }
         });
-        const [requestOptions] = this.requestSpy.firstCall.args;
+        const [, requestOptions] = this.requestSpy.firstCall.args;
         expect(requestOptions).to.have.property("headers").that.has.property("X-test", "test");
     });
 
     it("adds the trailing slash to the directory name if missing", async function () {
         await this.client.createDirectory("/subdirectory-1122");
-        const [requestOptions] = this.requestSpy.firstCall.args;
-        expect(requestOptions.url).to.satisfy(url => url.endsWith("/"));
+        const [url] = this.requestSpy.firstCall.args;
+        expect(url).to.satisfy((url: string) => url.endsWith("/"));
     });
 
     describe("with recursive option", function () {
         it("supports creating deep directories", async function () {
-            const newDir = path.resolve(__dirname, "../../testContents/a/b/c/d/e");
-            expect(directoryExists(newDir)).to.be.false;
+            const newDir = path.resolve(dirname, "../../testContents/a/b/c/d/e");
+            expect(directoryExists.sync(newDir)).to.be.false;
             await this.client.createDirectory("/a/b/c/d/e", { recursive: true });
-            expect(directoryExists(newDir)).to.be.true;
+            expect(directoryExists.sync(newDir)).to.be.true;
         });
 
         it("supports creating deep directories which partially exist", async function () {
-            const newDir = path.resolve(__dirname, "../../testContents/sub1/a/b");
-            expect(directoryExists(newDir)).to.be.false;
+            const newDir = path.resolve(dirname, "../../testContents/sub1/a/b");
+            expect(directoryExists.sync(newDir)).to.be.false;
             await this.client.createDirectory("/sub1/a/b", { recursive: true });
-            expect(directoryExists(newDir)).to.be.true;
+            expect(directoryExists.sync(newDir)).to.be.true;
         });
 
         it("has no effect when all paths exist", async function () {
