@@ -1,7 +1,10 @@
-const { AuthType } = require("../../../dist/node/index.js");
-const { createWebDAVClient } = require("../../helpers.node.js");
+import { expect } from "chai";
+import { AuthType } from "../../../source/index.js";
+import { createWebDAVClient } from "../../helpers.node.js";
 
-describe("getFileDownloadLink", function () {
+const CONTENT_TYPE_SUFFIX = "?Content-Type=application/octet-stream";
+
+describe("getFileUploadLink", function () {
     beforeEach(function () {
         this.clientPub = createWebDAVClient("http://test.com");
         this.clientAuth = createWebDAVClient("https://test.com", {
@@ -11,13 +14,13 @@ describe("getFileDownloadLink", function () {
     });
 
     it("generates authenticated links", function () {
-        const link = this.clientAuth.getFileDownloadLink("/test/file.txt");
-        expect(link).to.equal("https://user:pass@test.com/test/file.txt");
+        const link = this.clientAuth.getFileUploadLink("/test/file.txt");
+        expect(link).to.equal(`https://user:pass@test.com/test/file.txt${CONTENT_TYPE_SUFFIX}`);
     });
 
     it("generates unauthenticated links", function () {
-        const link = this.clientPub.getFileDownloadLink("/test/file.txt");
-        expect(link).to.equal("http://test.com/test/file.txt");
+        const link = this.clientPub.getFileUploadLink("/test/file.txt");
+        expect(link).to.equal(`http://test.com/test/file.txt${CONTENT_TYPE_SUFFIX}`);
     });
 
     it("throws for digest authentication", function () {
@@ -27,7 +30,7 @@ describe("getFileDownloadLink", function () {
             authType: AuthType.Digest
         });
         expect(() => {
-            this.client.getFileDownloadLink("/test/file.txt");
+            this.client.getFileUploadLink("/test/file.txt");
         }).to.throw(/Unsupported auth type.+digest/i);
     });
 });
