@@ -1,16 +1,20 @@
-const path = require("path");
-const fileExists = require("exists-file").sync;
-const directoryExists = require("directory-exists").sync;
-const {
+import path from "path";
+import { fileURLToPath } from "url";
+import fileExists from "exists-file";
+import directoryExists from "directory-exists";
+import { expect } from "chai";
+import {
     SERVER_PASSWORD,
     SERVER_PORT,
     SERVER_USERNAME,
     clean,
     createWebDAVClient,
     createWebDAVServer
-} = require("../../helpers.node.js");
+} from "../../helpers.node.js";
 
-const TEST_CONTENTS = path.resolve(__dirname, "../../testContents");
+const dirname = path.dirname(fileURLToPath(import.meta.url));
+
+const TEST_CONTENTS = path.resolve(dirname, "../../testContents");
 
 describe("moveFile", function () {
     beforeEach(function () {
@@ -29,22 +33,22 @@ describe("moveFile", function () {
 
     it("moves files from one directory to another", function () {
         return this.client.moveFile("/alrighty.jpg", "/sub1/alrighty.jpg").then(function () {
-            expect(fileExists(path.join(TEST_CONTENTS, "./alrighty.jpg"))).to.be.false;
-            expect(fileExists(path.join(TEST_CONTENTS, "./sub1/alrighty.jpg"))).to.be.true;
+            expect(fileExists.sync(path.join(TEST_CONTENTS, "./alrighty.jpg"))).to.be.false;
+            expect(fileExists.sync(path.join(TEST_CONTENTS, "./sub1/alrighty.jpg"))).to.be.true;
         });
     });
 
     it("moves directories from one directory to another", function () {
         return this.client.moveFile("/webdav", "/sub1/webdav").then(function () {
-            expect(directoryExists(path.join(TEST_CONTENTS, "./webdav"))).to.be.false;
-            expect(directoryExists(path.join(TEST_CONTENTS, "./sub1/webdav"))).to.be.true;
+            expect(directoryExists.sync(path.join(TEST_CONTENTS, "./webdav"))).to.be.false;
+            expect(directoryExists.sync(path.join(TEST_CONTENTS, "./sub1/webdav"))).to.be.true;
         });
     });
 
     it("moves files from one name to another", function () {
         return this.client.moveFile("/alrighty.jpg", "/renamed.jpg").then(function () {
-            expect(fileExists(path.join(TEST_CONTENTS, "./alrighty.jpg"))).to.be.false;
-            expect(fileExists(path.join(TEST_CONTENTS, "./renamed.jpg"))).to.be.true;
+            expect(fileExists.sync(path.join(TEST_CONTENTS, "./alrighty.jpg"))).to.be.false;
+            expect(fileExists.sync(path.join(TEST_CONTENTS, "./renamed.jpg"))).to.be.true;
         });
     });
 });
