@@ -1,4 +1,4 @@
-import { AuthHeader, OAuthToken } from "../types";
+import { AuthHeader, OAuthToken } from "../types.js";
 
 export function generateTokenAuthHeader(token: OAuthToken): AuthHeader {
     return `${token.token_type} ${token.access_token}`;
