@@ -1,5 +1,5 @@
-export { createClient } from "./factory";
-export { getPatcher } from "./compat/patcher";
-export * from "./types";
+export { createClient } from "./factory.js";
+export { getPatcher } from "./compat/patcher.js";
+export * from "./types.js";
 
-export { parseStat, parseXML } from "./tools/dav";
+export { parseStat, parseXML } from "./tools/dav.js";
