@@ -1,5 +1,5 @@
-import { getStat } from "./stat";
-import { WebDAVClientContext, WebDAVMethodOptions } from "../types";
+import { getStat } from "./stat.js";
+import { WebDAVClientContext, WebDAVMethodOptions } from "../types.js";
 
 export async function exists(
     context: WebDAVClientContext,
