@@ -1,21 +1,26 @@
-import { dirname } from "path-posix";
+import { Layerr } from "layerr";
+import path from "path-posix";
 
 const SEP_PATH_POSIX = "__PATH_SEPARATOR_POSIX__";
 const SEP_PATH_WINDOWS = "__PATH_SEPARATOR_WINDOWS__";
 
-export function encodePath(path) {
-    const replaced = path.replace(/\//g, SEP_PATH_POSIX).replace(/\\\\/g, SEP_PATH_WINDOWS);
-    const formatted = encodeURIComponent(replaced);
-    return formatted.split(SEP_PATH_WINDOWS).join("\\\\").split(SEP_PATH_POSIX).join("/");
+export function encodePath(filePath: string): string {
+    try {
+        const replaced = filePath.replace(/\//g, SEP_PATH_POSIX).replace(/\\\\/g, SEP_PATH_WINDOWS);
+        const formatted = encodeURIComponent(replaced);
+        return formatted.split(SEP_PATH_WINDOWS).join("\\\\").split(SEP_PATH_POSIX).join("/");
+    } catch (err) {
+        throw new Layerr(err, "Failed encoding path");
+    }
 }
 
-export function getAllDirectories(path: string): Array<string> {
-    if (!path || path === "/") return [];
-    let currentPath = path;
+export function getAllDirectories(directory: string): Array<string> {
+    if (!directory || directory === "/") return [];
+    let currentPath = directory;
     const output: Array<string> = [];
     do {
         output.push(currentPath);
-        currentPath = dirname(currentPath);
+        currentPath = path.dirname(currentPath);
     } while (currentPath && currentPath !== "/");
     return output;
 }
